+++ src/choices-guide/component/edit/form.jsx
@@ -18,7 +18,7 @@ export default class OpenStadComponentChoicesGuideForm extends OpenStadComponent
     super(props, {
       siteId: null,
       loginUrl: null,
-      noOfQuestionsToShow: 1,
+      noOfQuestionsToShow: 100,
       api: {
         url: null
       },